+++ src/network/AhoyWifiEsp8266.h
@@ -23,6 +23,7 @@ class AhoyWifi : public AhoyNetwork {
         }
 
         void tickNetworkLoop() override {
+            AhoyNetwork::tickNetworkLoop();
             if(mAp.isEnabled())
                 mAp.tickLoop();